+++ rpnpy/__init__.py
@@ -9,7 +9,7 @@ if sys.version_info < (3,):
 # will not be found by the version() function in ../setup.py
 #
 # Remember to update ../CHANGELOG.md describing what's new in each version.
-__version__ = '1.0.24'
+__version__ = '1.0.25'
 
 # Keep Python linters quiet.
 _ = Calculator